Build for windows
